Move compose/procfile into existing manifest package
