+++ internal/core/server/workspace.go
@@ -12,10 +12,10 @@ import (
 	"github.com/deref/exo/internal/chrono"
 	"github.com/deref/exo/internal/core/api"
 	state "github.com/deref/exo/internal/core/state/api"
-	"github.com/deref/exo/internal/external/procfile"
 	"github.com/deref/exo/internal/gensym"
 	logd "github.com/deref/exo/internal/logd/api"
 	"github.com/deref/exo/internal/manifest"
+	"github.com/deref/exo/internal/manifest/procfile"
 	"github.com/deref/exo/internal/providers/core"
 	"github.com/deref/exo/internal/providers/core/components/invalid"
 	"github.com/deref/exo/internal/providers/core/components/log"
