+++ internal/manifest/compose/parse.go
